# src/modules/permission/permission.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Patch,
  Param,
  Delete,
  UseGuards,
} from "@nestjs/common";
import { PermissionService } from "./permission.service";
import { CreatePermissionDto } from "./dto/create-permission.dto";
import { UpdatePermissionDto } from "./dto/update-permission.dto";
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBearerAuth,
} from "@nestjs/swagger";
import { AuthGuard } from "@nestjs/passport";
import { Permissions } from "src/common/decorators/permissions.decorator";
import { PermissionGuard } from "src/common/guards/permission.guard";
import { Permission } from "./entities/permission.entity";

@Controller("permission")
@ApiTags("权限管理")
@UseGuards(AuthGuard("jwt"), PermissionGuard)
@ApiBearerAuth()
export class PermissionController {
  constructor(private readonly permissionService: PermissionService) {}

  @Post()
  @ApiOperation({ summary: "创建权限" })
  @ApiResponse({ status: 201, description: "创建成功", type: Permission })
  @ApiResponse({ status: 400, description: "请求参数错误" })
  @ApiResponse({ status: 401, description: "未授权" })
  @ApiResponse({ status: 403, description: "权限不足" })
  @Permissions("permission:create")
  create(@Body() createPermissionDto: CreatePermissionDto) {
    return this.permissionService.create(createPermissionDto);
  }

  @Get()
  @ApiOperation({ summary: "获取所有权限" })
  @ApiResponse({ status: 200, description: "获取成功", type: [Permission] })
  @ApiResponse({ status: 401, description: "未授权" })
  @ApiResponse({ status: 403, description: "权限不足" })
  @Permissions("permission:read")
  findAll() {
    return this.permissionService.findAll();
  }

  @Get(":id")
  @ApiOperation({ summary: "根据ID获取权限" })
  @ApiResponse({ status: 200, description: "获取成功", type: Permission })
  @ApiResponse({ status: 401, description: "未授权" })
  @ApiResponse({ status: 403, description: "权限不足" })
  @ApiResponse({ status: 404, description: "权限不存在" })
  @Permissions("permission:read")
  findOne(@Param("id") id: string) {
    return this.permissionService.findOne(+id);
  }

  @Patch(":id")
  @ApiOperation({ summary: "更新权限" })
  @ApiResponse({ status: 200, description: "更新成功", type: Permission })
  @ApiResponse({ status: 400, description: "请求参数错误" })
  @ApiResponse({ status: 401, description: "未授权" })
  @ApiResponse({ status: 403, description: "权限不足" })
  @ApiResponse({ status: 404, description: "权限不存在" })
  @Permissions("permission:update")
  update(
    @Param("id") id: string,
    @Body() updatePermissionDto: UpdatePermissionDto,
  ) {
    return this.permissionService.update(+id, updatePermissionDto);
  }

  @Delete(":id")
  @ApiOperation({ summary: "删除权限" })
  @ApiResponse({ status: 200, description: "删除成功" })
  @ApiResponse({ status: 401, description: "未授权" })
  @ApiResponse({ status: 403, description: "权限不足" })
  @ApiResponse({ status: 404, description: "权限不存在" })
  @Permissions("permission:delete")
  remove(@Param("id") id: string) {
    return this.permissionService.remove(+id);
  }
}