# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/permission": {
      "post": {
        "operationId": "PermissionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建权限",
        "tags": [
          "权限管理"
        ]
      },
      "get": {
        "operationId": "PermissionController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有权限",
        "tags": [
          "权限管理"
        ]
      }
    },
    "/permission/{id}": {
      "get": {
        "operationId": "PermissionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "权限不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "根据ID获取权限",
        "tags": [
          "权限管理"
        ]
      },
      "patch": {
        "operationId": "PermissionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "权限不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新权限",
        "tags": [
          "权限管理"
        ]
      },
      "delete": {
        "operationId": "PermissionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "权限不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除权限",
        "tags": [
          "权限管理"
        ]
      }
    },
...
    "/user": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建用户",
        "tags": [
          "用户管理"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "username",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取用户列表",
        "tags": [
          "用户管理"
        ]
      }
    },
...
    "/order/{id}": {
      "get": {
        "operationId": "OrderController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "404": {
            "description": "订单不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取订单详情",
        "tags": [
          "订单管理"
        ]
      }
    },
...
    "/payment/create": {
      "post": {
        "operationId": "PaymentController_createPayment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "支付创建成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建支付",
        "tags": [
          "支付管理"
        ]
      }
    },
    "/payment/notify/alipay": {
      "post": {
        "operationId": "PaymentController_alipayNotify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlipayNotifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "回调处理成功"
          }
        },
        "summary": "支付宝支付回调",
        "tags": [
          "支付管理"
        ]
      }
    },
    "/payment/notify/wechat": {
      "post": {
        "operationId": "PaymentController_wechatNotify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatNotifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "回调处理成功"
          }
        },
        "summary": "微信支付回调",
        "tags": [
          "支付管理"
        ]
      }
    },
    "/payment/record/{id}": {
      "get": {
        "operationId": "PaymentController_findPaymentRecord",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询支付记录",
        "tags": [
          "支付管理"
        ]
      }
    },
    "/payment/order/{orderId}": {
      "get": {
        "operationId": "PaymentController_findPaymentByOrderId",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询订单支付记录",
        "tags": [
          "支付管理"
        ]
      }
    },
    "/payment/user": {
      "get": {
        "operationId": "PaymentController_findUserPayments",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询用户支付记录",
        "tags": [
          "支付管理"
        ]
      }
    },
    "/payment/notify/epay": {
      "get": {
        "operationId": "PaymentController_epayNotify",
        "parameters": [
          {
            "name": "pid",
            "required": true,
            "in": "query",
            "description": "商户ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trade_no",
            "required": true,
            "in": "query",
            "description": "易支付订单号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "out_trade_no",
            "required": true,
            "in": "query",
            "description": "商户订单号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trade_status",
            "required": true,
            "in": "query",
            "description": "交易状态",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "money",
            "required": true,
            "in": "query",
            "description": "交易金额",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "description": "支付类型",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": true,
            "in": "query",
            "description": "商品名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sign",
            "required": true,
            "in": "query",
            "description": "签名",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sign_type",
            "required": true,
            "in": "query",
            "description": "签名类型",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "回调处理成功"
          }
        },
        "summary": "易支付支付回调",
        "tags": [
          "支付管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "PicArt API",
    "description": "PicArt 图片社区 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "权限名称",
            "example": "user:create"
          },
          "description": {
            "type": "string",
            "description": "权限描述",
            "example": "创建用户"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "Permission": {
        "type": "object",
        "properties": {}
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "权限名称",
            "example": "user:create"
          },
          "description": {
            "type": "string",
            "description": "权限描述",
            "example": "创建用户"
          }
        }
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "admin"
          },
          "nickname": {
            "type": "string",
            "description": "昵称",
            "example": "管理员"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "123456"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "admin@example.com"
          },
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138000"
          },
          "roleIds": {
            "description": "角色ID列表（仅超级管理员可指定）",
            "example": [
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "wallet": {
            "type": "number",
            "description": "钱包余额",
            "example": 0
          },
          "inviteCode": {
            "type": "string",
            "description": "邀请码",
            "example": "INV123456789"
          },
          "verificationCode": {
            "type": "string",
            "description": "邮箱验证码",
            "example": "123456"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "number",
            "description": "订单ID"
          },
          "paymentMethod": {
            "type": "string",
            "description": "支付方式",
            "enum": [
              "ALIPAY",
              "WECHAT",
              "BALANCE",
              "EPAY"
            ]
          },
          "returnUrl": {
            "type": "string",
            "description": "支付完成后的跳转地址"
          },
          "type": {
            "type": "string",
            "description": "支付类型（EPAY支付方式时必传）",
            "enum": [
              "wxpay",
              "alipay",
              "usdt"
            ]
          }
        },
        "required": [
          "orderId",
          "paymentMethod"
        ]
      },
      "AlipayNotifyDto": {
        "type": "object",
        "properties": {
          "trade_no": {
            "type": "string",
            "description": "支付宝交易号"
          },
          "out_trade_no": {
            "type": "string",
            "description": "商户订单号"
          },
          "trade_status": {
            "type": "string",
            "description": "交易状态"
          },
          "total_amount": {
            "type": "string",
            "description": "交易金额"
          },
          "buyer_id": {
            "type": "string",
            "description": "买家支付宝用户号"
          },
          "sign": {
            "type": "string",
            "description": "签名"
          },
          "sign_type": {
            "type": "string",
            "description": "签名类型"
          }
        },
        "required": [
          "trade_no",
          "out_trade_no",
          "trade_status",
          "total_amount",
          "buyer_id",
          "sign",
          "sign_type"
        ]
      },
      "WechatNotifyDto": {
        "type": "object",
        "properties": {
          "transaction_id": {
            "type": "string",
            "description": "微信支付订单号"
          },
          "out_trade_no": {
            "type": "string",
            "description": "商户订单号"
          },
          "trade_state": {
            "type": "string",
            "description": "交易状态"
          },
          "amount": {
            "type": "string",
            "description": "交易金额"
          },
          "openid": {
            "type": "string",
            "description": "用户标识"
          },
          "sign": {
            "type": "string",
            "description": "签名"
          }
        },
        "required": [
          "transaction_id",
          "out_trade_no",
          "trade_state",
          "amount",
          "openid",
          "sign"
        ]
      },
...
    }
  }
}